requirements.txt
# Core dependencies for Open Sales Stack MCP servers
crawl4ai[all]
mcp[cli]
python-dotenv
